refactor: Step 5 — reducer を src/hooks/card-shogi/reducer.ts に分離
use-card-shogi-game.ts は 1,070 行に肥大化しており、純粋な reducer
ロジックと React hook 副作用 (useEffect / useCallback / DB 永続化) が
混在していた。Step 5 (Issue #107) として reducer 関数本体 + 内部 helper
+ Action 型を src/hooks/card-shogi/reducer.ts に切り出し、フック側は
useReducer + useEffect + 公開 API のみを担う薄い shell に縮小。

- src/hooks/card-shogi/reducer.ts (新規 708 行):
  - ShogiAction / Action 型 (export)
  - CardShogiGameStateInternal 型 (export)
  - makeMoveWithEffects / isKingInCheckAfterMove (内部 helper)
  - reducer 関数本体 (export)
- use-card-shogi-game.ts: 1,070 → 311 行 (71% 縮減)
  - reducer に必要な imports (effects.ts / definitions.ts / board.ts /
    moves.ts / rules.ts) を新ファイル側に集約し、フック側は
    moveToNotation / getAiMove / saveCardShogiMove / isValidCardTargetSquare
    のみに削減

ロジックは 1 行も変更なし (move-only リファクタ)。export shape も
useCardShogiGame のみで不変。

Issue #107 Step 5
